refactor: move task into agent package, clean up redundant API
- pkg/task/ → pkg/agent/task/: task management is internal to the
  agent system, not a standalone concern

- Remove ClearOnComplete (unused), unexport killAll (only called by
  Shutdown), merge FormatCompletion/formatCompletion into one function

- fireCompletion: log panics instead of silently swallowing them

- RunSkill: replace variadic Option with Config.Base field, call
  Config.Run() directly instead of going through functional options
